updates to code.. but still no build fix
