+++ crates/holons_client/Cargo.toml
@@ -13,14 +13,18 @@ derive-new = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 anyhow = "1.0"
+tokio = { version = "1", features = ["full"] }
+
 
 # Holochain Dependencies
 hdi = { workspace = true }
 hdk = { workspace = true }
+holochain_conductor_api = "0.3"
+holochain_types = "0.3"
+holochain_client = "0.5"
+holochain_zome_types = "0.3"
+holochain_websocket = "0.3"
 
-# dependencies for the holochain client library (currently SweetConductor 
-holochain = { version = "0.3", default-features = false, features = ["test_utils", "sqlite-encrypted"] }
-tokio = { version = "1.35.1", features = ["full"] }
 
 
 # MAP Dependencies
