+++ crates/holons_client/src/zome_client.rs
@@ -1,71 +1,147 @@
 
-use dances_core::dance_request::DanceRequest;
-use dances_core::dance_response::{DanceResponse, ResponseStatusCode};
-//use hdi::prelude::AgentPubKey;
 
-use hdk::prelude::CellId;
-use holochain::conductor::api::error::ConductorApiError;
-use holochain::sweettest::{SweetAgents, SweetApp, SweetCell, SweetConductor, SweetDnaFile};
-use holons_core::core_shared_objects::HolonError;
 
+use holochain_client::{AdminWebsocket, AgentPubKey, AppWebsocket, AuthorizeSigningCredentialsPayload, ClientAgentSigner, ConductorApiError, InstallAppPayload, InstalledAppId};
+use holochain_conductor_api::CellInfo;
+use holochain_types::prelude::{CellId, ExternIO};
+use holochain_types::websocket::AllowedOrigins;
+use serde::{Deserialize, Serialize};
+use std::{io::Error, io::ErrorKind, net::Ipv4Addr};
+use std::{
+    collections::HashMap
+};
 
 pub trait ZomeClient: Sized {
-    /// install calls the install_app method with the APP_ID and HAPP_FILEPATH
-    fn install() -> impl std::future::Future<Output = Result<Self, ConductorApiError>> + Send;
-    /// install_app installs the app with the given app_name and happ_url
-    fn install_app(app_name:&str, happ_url:Option<&str>) -> impl std::future::Future<Output = Result<Self, ConductorApiError>> + Send;
-    /// zomecall makes a zome call to the given cell_id, zome_name, and fn_name with the given request
-    async fn zomecall(self, cell_id:CellId, zome_name:&str, fn_name:&str, request:DanceRequest) -> Result<DanceResponse, HolonError>;
-    
-    //TODO:  async fn wait_on_signal(&self, cell_id:CellId)-> Result<(), ConductorApiError>;
+ async fn init(app_id:String, admin_port:u16) -> Result<Self, ConductorApiError>;
+ async fn zomecall(self, cell_id:CellId, zome_name:&str, fn_name:&str, payload:ExternIO) -> Result<ExternIO, ConductorApiError>;
+ fn get_cell_id_by_role(&self, role: Option<&str>) -> Result<CellId,ConductorApiError>;
+ //async fn wait_on_signal(&self, cell_id:CellId)-> Result<(), ConductorApiError>;
 }
 
-///The AppInstallation is created on Install and then available for use to make zome_calls / other commands
 #[derive(Debug)]
-pub struct AppInstallation {
-    pub conductor: SweetConductor,
-    pub app: SweetApp,
-    pub cells: Vec<SweetCell>,
+pub struct AppSessionData {
+    pub admin_port: u16,
+    pub app_id: InstalledAppId,
+    pub cells: HashMap<String,Vec<CellInfo>>,
     //pub signer: ClientAgentSigner,
 }
 
-const DNA_FILEPATH: &str = "../../workdir/map_holons.dna";
+const PORT: u16 = 9999; //this will be passed in from the launcher / sandbox - better the os chooses a port than hardcoding
 const APP_ID: &str = "map_holons";
-const HAPP_FILEPATH: &str = "../../workdir/map_holons.happ";
+const DNA_FILEPATH: &str = "../fixture/test.dna";
+const HAPP_PATH: &str = "workdir/hello-world.happ";
 
 
-
-impl ZomeClient for AppInstallation {
-    async fn install() -> Result<AppInstallation, ConductorApiError> {
-        Self::install_app(APP_ID, Some(HAPP_FILEPATH)).await
-    }
-    async fn install_app(app_id:&str, _happ_url:Option<&str>) -> Result<AppInstallation, ConductorApiError> {
-
-        let dna = SweetDnaFile::from_bundle(std::path::Path::new(&DNA_FILEPATH)).await.unwrap();
-        let mut conductor = SweetConductor::from_standard_config().await;
-        let holo_core_agent = SweetAgents::one(conductor.keystore()).await;
-        let app = conductor
-            .setup_app_for_agent(app_id, holo_core_agent.clone(), &[dna.clone()])
+impl ZomeClient for AppSessionData {
+    async fn init(app_id:String, admin_port:u16) -> Result<Self, ConductorApiError> {
+        // Connect admin web socket
+        let admin_ws = AdminWebsocket::connect((Ipv4Addr::LOCALHOST, admin_port))
             .await
-            .unwrap();
+            .map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(Error::new(ErrorKind::ConnectionRefused, arg0.to_string())))?;
 
-        let cells = &app.cells().clone();
-        Ok(Self{conductor, app, cells:cells.to_vec()})    
+        // Assume App is installed and enabled
+        let appdata = admin_ws.enable_app(app_id.clone()).await?;
+        let cell_data = appdata.app.cell_info;//.into_values().next().unwrap();
+        return Ok(Self{admin_port, app_id, cells: cell_data});
     }
 
-    async fn zomecall(self, cell_id:CellId, zome_name:&str, fn_name:&str, request:DanceRequest) -> Result<DanceResponse, HolonError> {
+    async fn zomecall(self, cell_id:CellId, zome_name:&str, fn_name:&str, payload:ExternIO) -> Result<ExternIO, ConductorApiError> {
+        // ******** SIGNED ZOME CALL  ********
 
-        let zome = self.conductor.get_sweet_cell(cell_id).map_err(|err: ConductorApiError|HolonError::WasmError(err.to_string()))?.zome(zome_name);
-        println!("{:?}", zome);
-        let response: DanceResponse = self.conductor.call::<DanceRequest,DanceResponse>(&zome, fn_name, request).await;
-        match response.status_code {
-            ResponseStatusCode::OK => return Ok(response),
-            ResponseStatusCode::Accepted => return Ok(response),
-            _ => return Err(HolonError::WasmError(response.status_code.to_string())),
-        };
+        // Connect admin web socket
+        let admin_ws = AdminWebsocket::connect((Ipv4Addr::LOCALHOST, self.admin_port))
+            .await
+            .map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(Error::new(ErrorKind::ConnectionRefused, arg0.to_string())))?;
+
+//            .map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(holochain_websocket::Error::new(ErrorKind::ConnectionRefused, (arg0.to_string()))))?;
+
+            // 0.5 .map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(holochain_websocket::WebsocketError::Other(arg0.to_string())))?;
+//admin_ws.attach_app_interface(port, allowed_origins, installed_app_id)
+        let credentials = admin_ws
+            .authorize_signing_credentials(AuthorizeSigningCredentialsPayload {
+                cell_id: cell_id.clone(),
+                functions: None,
+            })
+            .await
+            .map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(Error::new(ErrorKind::ConnectionRefused, arg0.to_string())))?;
+        let signer = ClientAgentSigner::default();
+        signer.add_credentials(cell_id.clone(), credentials);
+
+        // Connect app agent client
+        let app_ws_port = admin_ws
+          .attach_app_interface(0, AllowedOrigins::Any, None)
+          .await?;
+
+        let token_issued = admin_ws
+          .issue_app_auth_token(self.app_id.clone().into())
+          .await?;
+        let app_ws = AppWebsocket::connect(
+            (Ipv4Addr::LOCALHOST, app_ws_port),
+            token_issued.token,
+            signer.clone().into(),
+        )
+        .await
+        .map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(Error::new(ErrorKind::ConnectionRefused, (arg0.to_string()))))?;
+        //.map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(holochain_websocket::WebsocketError::Other(arg0.to_string())))?;
+
+        let response = app_ws
+            .call_zome(
+                cell_id.clone().into(),
+                zome_name.into(),
+                fn_name.into(),
+                payload
+                //.map_err(|arg0: anyhow::Error| ConductorApiError::WebsocketError(Error::new(ErrorKind::ConnectionRefused, arg0.to_string())))?;
+                //.map_err(|e| ConductorApiError::WebsocketError(Error::new(ErrorKind::ConnectionRefused, (e.to_string()))))?,
+                //.map_err(|e| ConductorApiError::WebsocketError(holochain_websocket::WebsocketError::Other(e.to_string())))?,
+            )
+            .await?;
+            println!("response: {:?}", ExternIO::decode::<String>(&response));
+            
+           // assert_eq!(
+           //     ExternIO::decode::<String>(&response).unwrap(),
+           //     fn_name.to_string()
+           // );
+        Ok(response)
     }
 
-   /* async fn wait_on_signal(&self, cell_id:CellId) -> Result<(), ConductorApiError> {
+    // Passing in None will return the first cell_id found, otherwise an error
+    fn get_cell_id_by_role(&self, role: Option<&str>) -> Result<CellId,ConductorApiError> {
+        if let Some(role) = role {
+            if let Some(cell_data) = self.cells.get(role) {
+                match cell_data[0].clone() {
+                    CellInfo::Provisioned(c) => Ok(c.cell_id),
+                    CellInfo::Cloned(c) => Ok(c.cell_id),
+                    _ => Err(ConductorApiError::CellNotFound)
+                }
+            } else {
+                Err(ConductorApiError::CellNotFound)
+            }
+        } else {
+            if let Some(cell_data) = self.cells.values().next().clone() {
+                match cell_data[0].clone() {
+                    CellInfo::Provisioned(c) => Ok(c.cell_id),
+                    CellInfo::Cloned(c) => Ok(c.cell_id),
+                    _ => Err(ConductorApiError::CellNotFound)
+                }
+            } else {
+                Err(ConductorApiError::CellNotFound)
+            }
+        }
+        
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct TestString(pub String);
+
+#[tokio::test(flavor = "multi_thread")]
+async fn mytest() {
+    let app: AppSessionData = AppSessionData::init(APP_ID.to_string(),PORT).await.unwrap();
+    let cell_id = app.get_cell_id_by_role(None).unwrap();
+    let payload = ExternIO::encode(TestString("commit".to_string())).unwrap();
+    app.zomecall(cell_id,"dances","dance",payload).await.unwrap();
+}
+    /*async fn wait_on_signal(&self, cell_id:CellId) -> Result<(), ConductorApiError> {
 
         // Connect admin client
         let admin_port = self.conductor.get_arbitrary_admin_websocket_port().unwrap();
@@ -118,30 +194,5 @@ impl ZomeClient for AppInstallation {
         barrier.wait();
             Ok(())
     }*/
-}
 
-// MOCK CONDUCTOR - here for reference - TODO remove
-
-/*pub async fn setup_conductor() -> (SweetConductor, AgentPubKey, SweetCell) {
-    let dna = SweetDnaFile::from_bundle(std::path::Path::new(&DNA_FILEPATH)).await.unwrap();
-
-    // let dna_path = std::env::current_dir().unwrap().join(DNA_FILEPATH);
-    // println!("{}", dna_path.to_string_lossy());
-    // let dna = SweetDnaFile::from_bundle(&dna_path).await.unwrap();
-
-    let mut conductor = SweetConductor::from_standard_config().await;
-
-    let holo_core_agent = SweetAgents::one(conductor.keystore()).await;
-    let app = conductor
-        .setup_app_for_agent("app", holo_core_agent.clone(), &[dna.clone()])
-        .await
-        .unwrap();
-
-    let cell = app.into_cells()[0].clone();
-
-    let agent_hash = holo_core_agent.into_inner();
-    let agent = AgentPubKey::from_raw_39(agent_hash).unwrap();
-
-    (conductor, agent, cell)
-}*/
